prometheus/requirements.txt
requests
bs4
redis
omniconf
python-dotenv
